refactor: extract match highlight logic to dedicated module
