+++ lua/grug-far.lua
@@ -159,73 +159,6 @@ local function setupBufRangeHighlight(buf, context)
   })
 end
 
---- sets up autocommands for current match highlight
----@param buf integer
----@param context grug.far.Context
----@private
-local function setupCurrentMatchHighlight(buf, context)
-  local resultsList = require('grug-far.render.resultsList')
-
-  -- re-apply highlight on entering grug buffer
-  vim.api.nvim_create_autocmd({ 'BufEnter' }, {
-    group = context.augroup,
-    buffer = buf,
-    callback = vim.schedule_wrap(function()
-      local location = resultsList.getResultLocationAtCursor(buf, context)
-      if not location then
-        return
-      end
-
-      local targetBuf = vim.fn.bufnr(location.filename)
-      if targetBuf == -1 or not vim.api.nvim_buf_is_valid(targetBuf) then
-        return
-      end
-
-      resultsList.highlightCurrentMatch(context, location, targetBuf)
-    end),
-  })
-
-  -- track cursor movement for live highlight
-  vim.api.nvim_create_autocmd({ 'CursorMoved' }, {
-    group = context.augroup,
-    buffer = buf,
-    callback = vim.schedule_wrap(function()
-      local location = resultsList.getResultLocationAtCursor(buf, context)
-      local new_key = location
-          and location.col
-          and (location.filename .. '|' .. location.lnum .. '|' .. location.col)
-        or nil
-
-      if new_key == context.state.currentMatchKey then
-        return
-      end
-
-      context.state.currentMatchKey = new_key
-
-      if not location then
-        resultsList.clearCurrentMatchHighlight(context)
-        return
-      end
-
-      local targetBuf = vim.fn.bufnr(location.filename)
-      if targetBuf == -1 or not vim.api.nvim_buf_is_valid(targetBuf) then
-        return
-      end
-
-      resultsList.highlightCurrentMatch(context, location, targetBuf)
-    end),
-  })
-
-  -- clear highlight on leaving grug buffer
-  vim.api.nvim_create_autocmd({ 'BufLeave' }, {
-    group = context.augroup,
-    buffer = buf,
-    callback = vim.schedule_wrap(function()
-      resultsList.clearCurrentMatchHighlight(context)
-    end),
-  })
-end
-
 ---@param context grug.far.Context
 ---@return integer windowId
 ---@private
@@ -388,7 +321,7 @@ function grug_far._open_internal(options, params)
     setupBufRangeHighlight(buf, context)
   end
 
-  setupCurrentMatchHighlight(buf, context)
+  require('grug-far.matchHighlight').setup(buf, context)
 
   require('grug-far.farBuffer').setupBuffer(win, buf, context, function()
     instance:_set_ready()
